refactor: asis 패키지에 있는 레거시 코드 삭제
