--- mvc/src/main/java/com/interface21/webmvc/servlet/mvc/tobe/handleradapter/ControllerHandlerAdapter.java
@@ -1,23 +1,0 @@
-package com.interface21.webmvc.servlet.mvc.tobe.handleradapter;
-
-import com.interface21.webmvc.servlet.ModelAndView;
-import com.interface21.webmvc.servlet.mvc.asis.Controller;
-import com.interface21.webmvc.servlet.view.JspView;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-
-public class ControllerHandlerAdapter implements HandlerAdapter {
-
-    @Override
-    public boolean supports(final Object handler) {
-        return (handler instanceof Controller);
-    }
-
-    @Override
-    public ModelAndView handle(final HttpServletRequest request,
-                               final HttpServletResponse response,
-                               final Object handler) throws Exception {
-        String viewName = ((Controller) handler).execute(request, response);
-        return new ModelAndView(new JspView(viewName));
-    }
-}
