+++ mvc/src/test/java/samples/ManualHandlerMapping.java
@@ -1,6 +1,5 @@
-package com.techcourse;
+package samples;
 
-import com.interface21.webmvc.servlet.mvc.asis.Controller;
 import com.interface21.webmvc.servlet.mvc.tobe.handlermapping.HandlerMapping;
 import jakarta.servlet.http.HttpServletRequest;
 import java.util.HashMap;
